requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.20
requests>=2.31.0
python-dotenv>=1.0.0

# Bilibili API and video processing
bilibili-api-python>=16.1.0
yt-dlp>=2023.12.0

# Video/Audio processing
moviepy>=1.0.3
opencv-python>=4.8.0
vosk>=0.3.45

# Text processing and NLP
jieba>=0.42.1
wordcloud>=1.9.3
numpy>=1.24.0
pillow>=10.0.0

# Data handling
pandas>=2.0.0

# HTTP client
httpx>=0.25.0
